requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.2 
pytorch==1.12.1
scikit-learn==1.2.0
scipy==1.9.3
torchvision==0.14.1


                 